refactor: separated gui files into smaller parts
