+++ src/gui.rs
@@ -1,12 +1,22 @@
 //! Gui driver
 use iced::{
-    button, executor, pane_grid, pick_list, scrollable, text_input,
+    executor, pane_grid,
     widget::{
         Button, Column, Container, PaneGrid, PickList, Row, Rule, Scrollable, Text, TextInput,
     },
     Align, Application, Clipboard, Command, Element, Length, Settings,
 };
-use serde::{Deserialize, Serialize};
+
+mod pane_content;
+mod perf_event;
+mod save_state;
+mod state;
+mod style;
+
+use pane_content::*;
+use perf_event::*;
+use save_state::*;
+use state::*;
 
 /// Run the Gui Launcher
 pub fn run_gui() -> iced::Result {
@@ -19,57 +29,6 @@ enum Gui {
     Loaded(State),
 }
 
-struct State {
-    tasks: Vec<Task>,
-    panes_state: pane_grid::State<Content>,
-    panes_created: usize,
-    data_pane: pane_grid::Pane,
-    log_pane: pane_grid::Pane,
-    task_pane: pane_grid::Pane,
-}
-
-/// Default state for Gui
-impl Default for State {
-    fn default() -> Self {
-        // First pane and first state is created here:
-        // task Pane, panes_state
-        let (mut panes_state, task_pane) = pane_grid::State::new(Content::new(PaneType::Task, 0));
-
-        // Second pane and first split is created here:
-        // data_pane, vert_split
-        let (data_pane, vert_split) = panes_state
-            .split(
-                pane_grid::Axis::Vertical,
-                &task_pane,
-                Content::new(PaneType::Main, 1),
-            )
-            .unwrap();
-
-        // Third plane and second split is created here:
-        // log_pane, horz_split
-        let (log_pane, horz_split) = panes_state
-            .split(
-                pane_grid::Axis::Horizontal,
-                &data_pane,
-                Content::new(PaneType::Log, 2),
-            )
-            .unwrap();
-
-        panes_state.resize(&vert_split, 0.17);
-        panes_state.resize(&horz_split, 0.88);
-
-        let tasks = Vec::new();
-
-        State {
-            tasks,
-            panes_state,
-            panes_created: 3,
-            data_pane,
-            task_pane,
-            log_pane,
-        }
-    }
-}
 /// Messages to be sent to the parent widget from
 /// other child widgets, and consumed on update
 #[derive(Debug, Clone)]
@@ -335,7 +294,7 @@ impl Application for Gui {
                         ),
                     })
                     .title_bar(title_bar)
-                    .style(style::Pane { is_focused: true })
+                    .style(style::widget::Pane { is_focused: true })
                 })
                 .width(Length::Fill)
                 .height(Length::Fill)
@@ -360,87 +319,6 @@ impl Application for Gui {
     }
 }
 
-/// Perf Commands to be used
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum PerfEvent {
-    Stat,
-    Record,
-    Report,
-    Annotate,
-    Top,
-    Bench,
-    Test,
-}
-
-/// Holds an enumerated array of PerfEvents
-impl PerfEvent {
-    const ALL: [PerfEvent; 7] = [
-        PerfEvent::Annotate,
-        PerfEvent::Bench,
-        PerfEvent::Record,
-        PerfEvent::Report,
-        PerfEvent::Stat,
-        PerfEvent::Test,
-        PerfEvent::Top,
-    ];
-}
-
-/// Default PerfEvent
-impl Default for PerfEvent {
-    fn default() -> PerfEvent {
-        PerfEvent::Test
-    }
-}
-
-/// Provide PerfEvents as String data types
-impl std::fmt::Display for PerfEvent {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            "{}",
-            match self {
-                PerfEvent::Annotate => "Annotate",
-                PerfEvent::Bench => "Bench",
-                PerfEvent::Record => "Record",
-                PerfEvent::Report => "Report",
-                PerfEvent::Stat => "Stat",
-                PerfEvent::Test => "Test",
-                PerfEvent::Top => "Top",
-            }
-        )
-    }
-}
-
-/// Widget Style
-mod style {
-    use iced::{container, Background, Color};
-
-    const SURFACE: Color = Color::from_rgb(
-        0xF2 as f32 / 255.0,
-        0xF3 as f32 / 255.0,
-        0xF5 as f32 / 255.0,
-    );
-
-    pub struct Pane {
-        pub is_focused: bool,
-    }
-
-    impl container::StyleSheet for Pane {
-        fn style(&self) -> container::Style {
-            container::Style {
-                background: Some(Background::Color(SURFACE)),
-                border_width: 2.0,
-                border_color: if self.is_focused {
-                    Color::BLACK
-                } else {
-                    Color::from_rgb(0.7, 0.7, 0.7)
-                },
-                ..Default::default()
-            }
-        }
-    }
-}
-
 /// Message to display while Gui is loading
 fn loading_message<'a>() -> Element<'a, Message> {
     Container::new(Text::new("Loading...").size(50))
@@ -448,184 +326,4 @@ fn loading_message<'a>() -> Element<'a, Message> {
         .height(Length::Fill)
         .center_y()
         .into()
-}
-
-/// Main pane Contexts
-enum Context {
-    Main,
-    NewEvent,
-}
-
-/// Pane Type
-enum PaneType {
-    Task,
-    Main,
-    Log,
-}
-
-/// States of all panes within the pane grid
-// every pane state must be held here
-struct Content {
-    input_value: String,
-    input: text_input::State,
-    selected_command: PerfEvent,
-    scroll: scrollable::State,
-    pick_list: pick_list::State<PerfEvent>,
-    id: usize,
-    data: String,
-    application: String,
-    pane_type: PaneType,
-    create_button: button::State,
-    launch_button: button::State,
-    context: Context,
-}
-
-/// Initialize pane states to default values
-impl Content {
-    fn new(pane_type: PaneType, id: usize) -> Self {
-        Content {
-            input_value: String::new(),
-            input: text_input::State::new(),
-            selected_command: PerfEvent::default(),
-            scroll: scrollable::State::new(),
-            pick_list: pick_list::State::default(),
-            pane_type,
-            id,
-            data: String::new(),
-            create_button: button::State::new(),
-            launch_button: button::State::new(),
-            application: String::new(),
-            context: Context::Main,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-// Currently running or previously ran events
-struct Task {
-    name: String,
-    application: String,
-    options: Vec<String>,
-}
-
-//customized from iced todo example.
-// source: https://github.com/hecrj/iced/blob/0.3/examples/todos/src/main.rs
-
-//Persistance
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct SavedState {
-    tasks: Vec<Task>,
-}
-
-#[derive(Debug, Clone)]
-/// Error type for load function
-enum LoadError {
-    FileError,
-    FormatError,
-}
-
-#[derive(Debug, Clone)]
-/// Error type for save function
-enum SaveError {
-    FileError,
-    WriteError,
-    FormatError,
-}
-
-#[cfg(not(target_arch = "wasm32"))]
-/// Saved state for Gui
-impl SavedState {
-    fn path() -> std::path::PathBuf {
-        let mut path = if let Some(project_dirs) =
-            directories_next::ProjectDirs::from("rs", "ruperf", "Tasks")
-        {
-            project_dirs.data_dir().into()
-        } else {
-            std::env::current_dir().unwrap_or(std::path::PathBuf::new())
-        };
-
-        path.push("tasks.json");
-
-        path
-    }
-
-    async fn load() -> Result<SavedState, LoadError> {
-        use async_std::prelude::*;
-
-        let mut contents = String::new();
-
-        let mut file = async_std::fs::File::open(Self::path())
-            .await
-            .map_err(|_| LoadError::FileError)?;
-
-        file.read_to_string(&mut contents)
-            .await
-            .map_err(|_| LoadError::FileError)?;
-
-        serde_json::from_str(&contents).map_err(|_| LoadError::FormatError)
-    }
-
-    async fn save(self) -> Result<(), SaveError> {
-        use async_std::prelude::*;
-
-        let json = serde_json::to_string_pretty(&self).map_err(|_| SaveError::FormatError)?;
-
-        let path = Self::path();
-
-        if let Some(dir) = path.parent() {
-            async_std::fs::create_dir_all(dir)
-                .await
-                .map_err(|_| SaveError::FileError)?;
-        }
-
-        {
-            let mut file = async_std::fs::File::create(path)
-                .await
-                .map_err(|_| SaveError::FileError)?;
-
-            file.write_all(json.as_bytes())
-                .await
-                .map_err(|_| SaveError::WriteError)?;
-        }
-
-        // This is a simple way to save at most once every couple seconds
-        async_std::task::sleep(std::time::Duration::from_secs(2)).await;
-
-        Ok(())
-    }
-}
-
-#[cfg(target_arch = "wasm32")]
-// Saved state for Gui (wasm32)
-impl SavedState {
-    fn storage() -> Option<web_sys::Storage> {
-        let window = web_sys::window()?;
-
-        window.local_storage().ok()?
-    }
-
-    async fn load() -> Result<SavedState, LoadError> {
-        let storage = Self::storage().ok_or(LoadError::FileError)?;
-
-        let contents = storage
-            .get_item("state")
-            .map_err(|_| LoadError::FileError)?
-            .ok_or(LoadError::FileError)?;
-
-        serde_json::from_str(&contents).map_err(|_| LoadError::FormatError)
-    }
-
-    async fn save(self) -> Result<(), SaveError> {
-        let storage = Self::storage().ok_or(SaveError::FileError)?;
-
-        let json = serde_json::to_string_pretty(&self).map_err(|_| SaveError::FormatError)?;
-
-        storage
-            .set_item("state", &json)
-            .map_err(|_| SaveError::WriteError)?;
-
-        let _ = wasm_timer::Delay::new(std::time::Duration::from_secs(2)).await;
-
-        Ok(())
-    }
 }
